符号表
[add](VoidTy现在用于指示无返回值的函数调用返回值，因此其不占用物理空间，在计算局部变量大小时其大小为0)
[add](部分后端符号表的代码，正在考虑其是否必要)

+++ backend/inc/variable_map.h
@@ -146,7 +146,7 @@ void print_single_info_map(List* this,int order,bool print_info);
 #define HashMap_foreach(map,k,v)  \
     HashMapFirst(map);                  \
     for(Pair* pair_ptr = HashMapNext(map);  \
-        (pair_ptr != NULL) && (k = pair_ptr->key) && (v = pair_ptr->value);    \
+        (pair_ptr != NULL) && (k = pair_ptr->key) && (v = (uint64_t)(pair_ptr->value));    \
         pair_ptr = HashMapNext(map))
